Temporary fix to allow target discovery during MPIPL
Techically we should not be discovering targets all over again,
we want the exact same thing that was given from the first IPL
(IE ignore and deconfigs that happened between). In order to run
MPIPL on a standalone model we need another mechanism to get the
targeting data. That mechanism doesn't exist yet so we will just
do another discovery for now.

Change-Id: I8193534f74c4f3fc2dcc795b705b5184e702c60d
RTC: 157651
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/30425
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>