Retrieve HW Keys' Hash from SBE Image on SBE Seeprom, if necessary
When not being updated via a Key Transition IPL, each SBE Seeprom's
HW Keys' Hash is stored in the SBE Version Structure on the SBE
Seeprom.  However, if that structure is old or becomes corrupted,
the HW Keys' Hash will be retrieved from the SBE Image that resides
on the SBE Seeprom itself.

Change-Id: I5f58a9adf5267aaff0332284228147d4fdf0e022
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/30596
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Stephen M. Cprek <smcprek@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>