Add infrastructure for sending setSystemConfig cmd to SBE
The SBE needs to know about other PROCs in the sytem. To do
this we are sending a 64 bit string that encodes the system
configuration down to the SBE at the end of the IPL. This
commit just sets up some of the infrastructure that will be needed
to send the message.

Change-Id: I1a4dfb4e54e043d94697fa809acf1eee8ba9f726
RTC: 160666
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/31069
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Matthew A. Ploetz <maploetz@us.ibm.com>