Update attribute/target xml with info from hb_temp_defaults
We have let hb_temp_defaults get a little out of hand and I hit
something that was blocked by the attributes there not being avail
on a target. Rather than just fix my one problem I decided to clean
house a little bit. This commit should make it so that we can remove
all of the attributes from hb_temp_defaults on the EKB side

Change-Id: Ie2728972a250545549714f4b2fc234417194db94
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/31737
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <v2cibmd@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>