Added error count mode 2 to MCBIST lab
MCBIST has three modes for debugging 0-2.  Mode 2 is the
most similar to IBM's prior version of MCBIST. It will also
be used as a starting point for the shmoo code.  The print
functionality is currently a starting point and very basic.
Only one mode is in this commit to keep the commit size small
for ease of review.

Change-Id: I4849aae16d3a2607477d24ee6e388823f3f77baa
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/32005
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Brian R. Silver <bsilver@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/32043
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>