Read/Write from pnor for R/W attribute section in Attrrp in MPIPL
During an MPIPL the targeting and attribute data in pnor cannot be
trusted so we use a copy of the info that we had stashed in memory
during runtime. We do want to perserve writes to this info and
if we only write it to memory it will not be preserved. To get
around this we are hooking up the R/W section to pnor during MPIPL

Change-Id: I5f98f5e5a522d923cc812060a5f8c959b74e683c
RTC:164480
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/32557
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>