PRD: fixed error handling for getSlaveRanks()/getMasterRanks()
getSlaveRanks() calls getMasterRanks() for each DIMM select, which
will assert if it does not find any master ranks on the DIMM select.
The issue is that it is possible for only one of the DIMM selects to
have configured ranks. So the code will assert when getSlaveRanks()
iterates on the non-configured DIMM select.

Change-Id: I233a5cbf6ac8df2f64a0ab4ea153e3d0d93bfc4a
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/33612
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Benjamin J. Weisenbeck <bweisenb@us.ibm.com>
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/33643
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>