PRD: Data Storage exception in BitString class
We found a bug where a BitString object was used on an area of memory
that was not CPU_WORD aligned. This memory was also allocated at the
very end of a page and the next page was not accessible by Hostboot. So
as when the last CPU_WORD of the data was accessed it crossed over into
the unaccessible page and cause and exception.

Change-Id: I71e87aeb4feaea1ea977d3b396674073a3d03e7b
CQ: SW352432
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/35011
Reviewed-by: Brian J. Stegmiller <bjs@us.ibm.com>
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/35012
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>