Quick fix to get boot side before SBE is started
Code can attempt to get the boot side for an SBE before the SBE is
started and the SCOM can be read.  In this case the boot side for the
master processor should be used.  First step of this fix is to remove
the #if logic around the block to select the master processor target.

Change-Id: Idb08165a3e2c9217adc17592ffb56b9434182acd
Depends-on: I8111a40c8aa53e1699b0381440584db3dd6c3fc6
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/35730
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>