IPL: Error log cleanup: BC8A2210 - SBE_PERM_SIDE_DIRTY_BAD_PATH
Long-term fix to get boot side before SBE is started
Code can attempt to get the boot side for an SBE before the SBE is
started and the SCOM can be read. In this case the boot side for the
master processor should be used. Second step of this fix is to
read the SCOM for the master processor target when finding the boot
side of the master processor SBE or a slave processor SBE that is not
started and read the SCOM for the slave processor target only after
its SBE is started.  A new processor target attribute will be set
after the processor's SBE has been started.

Change-Id: I12f97435f8872e41f06dcff6120abe4c24a1bb22
CQ: SW378865
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/35938
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>