Updates PHY to always terminate DQ/DQS
This update updates bit 49 of DP16_RD_DIA_CONFIG5 to be a 1.  This
cause the PHY to always terminate the DQ/DQS lines unless conducting
a write. The reasoning behind this update is to ensure that the
DQ/DQS voltage levels never drop to 0V as this could creates issues
in the PHY circuitry.  This update will need to be left on for
all planned design revisions of the chip.

Change-Id: I88c884e3f62342aa716855c4b950c81ebdc9b226
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/36238
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: ANDRE A. MARIN <aamarin@us.ibm.com>
Reviewed-by: JACOB L. HARVEY <jlharvey@us.ibm.com>
Dev-Ready: ANDRE A. MARIN <aamarin@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/36282
Reviewed-by: Hostboot Team <hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>