Remove enumeration types from FAPI attributes
Enumerated attributes inside of TARGETING are always 4 bytes long.
FAPI attributes can be of arbitrary length, even if enumerations
are specified.  This means that we have a size mismatch if we use
an enumerated attribute as the direct map to a fapi attribute.
Instead, we should always use the same basic type that matches
the FAPI attribute, but still declare the enumeration so it can
be used for MRW settings.

Added an explicit failure in xmltohb.pl that will prevent us
from making this mistake again.

Change-Id: Iab855e0e93d2d56922175a1d6cb165eed4d665fa
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/36240
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>