Remove deprecated SBE side resolution
The HB Bootloader now owns the responsibility of selecting the
appropriate side of PNOR to use when loading the HBB image.  There
is no longer a HBB pointer in the SBE image itself.  This commit
removes the code that dealt with that support.  There is also
some cleaning up of a few other related functions.

Change-Id: I35bfccb3590f81867d8222333cb4be66718828fa
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/36358
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>