Update quad power off so HB can call it on Slave Quads
Previously this HWP was only being called by the SBE down the MPIPL
but we also need to call it during HB to power down the slave quads.
There is a tricky workaround in this HWP where we need to save off
some ring data EQ pointers for the PB. On the SBE we just saved
the rings in global variables but we cannot do that in HB because
the Hcode would not have access to the global varibles. Instead
we will write the ring data to the OCC SRAM to do this I will
just pass the data out of the hwp via an out param
The same ring data is fetch during cache_initf procedure to init

Change-Id: Ife475cf89a7d24bdc595b8d8ea9a60aed4b4b8fb
Original-Change-Id: I4f20cd8dd9ca1b53604db1bc3020c0d57bbd1172
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/36828
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt K. Light <mklight@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>
Reviewed-by: YUE DU <daviddu@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>