Fixing XBUS PG lookup to mark both E44D and E55D as valid entries
Currently, there are two version of the MVPD PG keyword:
- E44D == XBUS0 bad
- E45D and E55D == XBUS0 and 2 bad
Spec indicates that both iox (second nibble) and pbiox
(third nibble) are bad for sforza and monza type modules.
We support generically the following cases:
- 0xE50D --> xbus chiplet good
- 0xE40D --> xbus chiplet good
and rely solely on the pbiox as the Xbus target indicator
(0x0040, 0x0020, 0x0010) for all types of chips.

Change-Id: If7a3c128180321adab083f3e40f2dd0a645318e7
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/37007
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>