move all non-volatile/openpower attributes from _hb to _openpower files
When pushing hb's attr/target type files to common-mrw-xml repo, we add
tags to represent which attributes are for openpower (bmc based system)
and which are for fsp based systems. All attributes in _hb files are
considered common for both platforms and do not have any tag. All
attributes in _openpower will get openpower tag and be only shown
to a system engineer if they are building a bmc based system.
Therefore, the source of the attribute file is important from now on.

Change-Id: I213348f4c4f372b0610b07ad3453f08f75d02224
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/38161
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew A. Ploetz <maploetz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>