Updating L3 LCO watermarks for HW406803
In DD1 there is a bug in how the LCO home region is calculated so
that the wrong bit44 vaule might be used. In the long run this
works out to about half of the LCO home region being dropped.
To help mitigate this, we're halfing watermarks for LCO's.

Change-Id: I3a38b06bd4520a24b8634facaa732baa5e008ecd
Original-Change-Id: I0d64d0e4a2bb72e1a413dc03612fc6e669b56be5
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/38874
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: ALEXANDER M. TAFT <amtaft@us.ibm.com>
Reviewed-by: Matt K. Light <mklight@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55489
CI-Ready: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M. Crowell <dcrowell@us.ibm.com>