Disable resetting on INT unit during MPIPL
Previously we were trying to pull the thead context before clearing
out the phs_thread_enabled regs when we initialized the intrrp
during mpipl. However I found that attempting this is causing a
recoverable fir which down the road creates a predictive gard
record. We don't need to be calling this in DD1. Not sure for DD2
yet but I have a story to revist

Change-Id: Ibb7d11463e80285354fda1355105cb58e2101ea7
RTC:172905
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/39344
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>