Disable cp_me from the L3 for Nimbus DD1 and DD2.0.
By default the L3 only sends these when the addres is non-local.
This is needed for the MCU to send early data. However MCU early
data is broken in DD1 and DD2.0 and this eats up X-link bandwidth,
so for DD1 and DD2.0 we're disabling cp_me from the L3.

Change-Id: I13a894f91434cc19f5f9502f54cc59bad9a0ae0f
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/39483
Reviewed-by: ALEXANDER M. TAFT <amtaft@us.ibm.com>
Reviewed-by: Joseph J. McGill <jmcgill@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt K. Light <mklight@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N. Tran <thi@us.ibm.com>