lpc: Add support for large read/write ops
On P9 the LPC is memory mapped. The LPC FW space used to map
the flash is generally acccessed to read or write rather large
amounts of data.

Its inefficient to go through the whole dispatch and locking for
every single 4 bytes read or written.

This adds the ability to request reads or writes of larger
quantities for FW space. The implementation uses memcpy whose
current implementation in HostBoot will perform well for 8
bytes aligned accesses, but will downgrate to bytes
accesses otherwise.

Change-Id: I9770f22da99d1e1b917f4ba2101d459483f1dee1
Signed-off-by: Benjamin Herrenschmidt <bherren@au1.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/39386
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>