ProcessMRW changes for dynamic i2c devices
HDAT has a requirement from hostboot to provide I2C information
from the connections in the mrw.  This change pulls the changes
out of the mrw and stores them under the master proc target.
Each index in the arrays created is a separate connection.
The information is then pulled out in i2c.C getDeviceInfo()

Change-Id: I725278bc482131a65b8cfb565e08ff0a37f4cb30
RTC:165485
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/39737
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>