ProcessMRW changes for dynamic i2c devices
HDAT has a requirement from hostboot to provide I2C information
from the connections in the mrw.  This change pulls the changes
out of the mrw and stores them under the master proc target.
Each index in the arrays created is a separate connection.
The information is then pulled out in i2c.C getDeviceInfo()

Change-Id: I378ef6520dc6d32ef623fd438e73881ce928b37d
RTC:165485
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42363
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>