Handle lack of overrides in regular RHB mode
A workaround was added to the code to allow attribute override
support on systems that are still using the old style of HBRT
data (not Reserved Hostboot Memory).  That code caused an error
to be thrown on systems that did use RHB.  The code was changed
to delete the error in this case insted of committing it.

Change-Id: I26e44965fbb0bf5ca73ca6d3e10feb01db28565d
CQ: SW386889
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/39941
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>