Update the interrupt routing for the Slave proc
During Hostboot we route all interrupts to the master proc.
During the MPIPL we send a PSU chip op to the SBE, if the slave
isnt set up to route to the master proc then we get a xstop.
This fix makes sure we update the slave proc's interrupt routing
on the MPIPL path

Change-Id: I41e2728e049128e631537fe016d4402264739f23
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40181
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>