Use effective fabric position for LPC operations in the bootloader
When the LPC BAR value is supplied in the SBE to BL structure, use it as
the base LPC address, otherwise, still use the LPC constant.  Save the
value being used in the Bootloader data (which is reorganized to align
data blocks on boundaries).
Also add hb-bldata to parse out the Bootloader data.

Change-Id: I1db19467464b90e0190c4df5f7404624c9423eb5
RTC: 173525
Depends-on: I2b0d1959c303df8c9c28c8f0a5b5be1e77aa154f
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40217
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>