Save space in Bootloader with exception handlers
The exception handlers save and restore the register used to save the
exception address.  The Bootloader does not need to preserve all
registers when an exception occurs and can remove two assembly
instructions per exception.  It will use gpr17 to save the exception
address, since that register is less used in Bootloader than gpr1.

Change-Id: Ia736040e3975508f40c106d4a8ea1a05f3b75c29
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40284
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>