bootloader: Remove hbi_ImageId
None of the bootloader code references the hbi_ImageId but it is
included anyway. If the version string is changed in the build sequence
this results in a new hbbi.bin image. When booting with this new image,
an SBE update will be triggered since the hbbl.bin is different than the
one in the on chip SBE.

Change-Id: Ia1f68dbd61310b707a7712f88e6c41cd9a670fbe
Signed-off-by: William A. Kennington III <wak@google.com>