bootloader: Remove hbi_ImageId
None of the bootloader code references the hbi_ImageId but it is
included anyway. If the version string is changed in the build sequence
this results in a new hbbl.bin image. When booting with this new image,
an SBE update will be triggered since the hbbl.bin is different than the
one in the on chip SBE.

Resolves #102

Signed-off-by: William A. Kennington III <wak@google.com>
Change-Id: I5c19408473bb68ee89ef864517b6964844aad9bd
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40822
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Stephen M. Cprek <smcprek@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>