Map BAR attributes based on data from Bootloader
If the master processor has no memory behind it the entire
memory map must be modified.  Each processor has its own statically
defined map that covers both memory and MMIOs.  If the master
has no memory, its memory map is swapped with another processor.
Each processor gets a new effective fabric id that is then used
to compute all of the BAR values for those processors.

The SBE boots with a certain memory map programmed into the master
processor.  That value is then passed up through the bootloader
into Hostboot.  This value is compared to the BAR values that
Hostboot assumes it is using.  Based on that comparison, various
attributes are computed to match the effective fabric positions.

Change-Id: I2b0d1959c303df8c9c28c8f0a5b5be1e77aa154f
RTC: 173528
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40359
Tested-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>