Bug fixes in hb script
sh uses '.' to source a file. For some reason 'source' is allowed, but
not on all systems and therefore is not portable.

The '==' comparison is only for strings. The '-eq' comparison should be
used for numeric comparisons.

'exit' takes a value of 0-255. Use of a negative number, such as '-1',
is will throw an error.

Change-Id: Ib617f632a9c9ad11a28ad0e07154f364d8572f57
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40579
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Elizabeth K. Liner <eliner@us.ibm.com>
Reviewed-by: Benjamin J. Weisenbeck <bweisenb@us.ibm.com>
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Brian J. Stegmiller <bjs@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>