sbe_extract_rc_handler finite state machine details (3/4)
This commits sets up the specific details for every function
included in the finite state machine.  Basically, it calls
a retry of the SBE if applicable, and then handles the error
or pass for what it needs to do next.  In the end, we either
pass with a working SBE or we fail, gard out the proc and continue.

Change-Id: I298e43e85bfd4f0591aefe289d6792f001fb24b2
RTC:173809
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/41075
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>