Reapply "Use query_cache_state hwp before scoming ex's in cme_firinit"
we cannot rely on the HWAS state because during an MPIPL the cores
get stopped and the SP doesnt know until an attr sync occurs w/
the platform. We must use the query_cache_state to safely to
determine if we can scom the ex targets. This had to be removed
because we were incorrectly querying the core EQ state.

Change-Id: I248e756b7e305aae0abe1714db610f1e69f2b066
CQ: SW388687
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40926
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Gregory S. Still <stillgs@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>