Reapply "Use query_cache_access_state to verify EQ scomable in.."
In corequad_init we cannot rely on hwas states because targets are
not always scommable even though they are functional. This adds a
check to ensure a ex is scommable before we attempt to scom it.
This had to be removed because we were incorrectly querying the
core EQ state. That issue is now resolved

Change-Id: I4a3c516b5063e1a898f2ebe7f0da68cfb965b737
CQ: SW388687
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/40927
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: Gregory S. Still <stillgs@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>