Verify the correct HW Key Hash is used in Customized SBE Image
This commit keeps track of what HW Key Hash is added to HBBL before
customizing the SBE Image and then checks that this HW Key Hash is found
in the customized SBE Image.  It did this by updating getHwKeyHashFromSbeImage()
to possibly read the HW Key Hash from system memory along with its default
behavior of reading it from a SBE Seeprom.

Change-Id: I0139fb959102de74b12874f30e7d2ec0bf918e3f
RTC:175330
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/41453
Reviewed-by: Stephen M. Cprek <smcprek@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>