Replaced calls to bmc for sensor type with static information.
The sendErrLogToBMC function was calling to the bmc in order
to determine the type of a sensor based upon a sensor number.
It was determined that the sensor type can be obtained from the
sensor number by data loaded during initialization via the
TARGETING namespace. The ipmi call from SenorBase::getSensorType
was removed from the sendErrLogToBMC function and replaced with
a call to IpmiConfigLookup::getSensorType. The getSensorType
function from the new IpmiConfigLookup class retrieves sensor
data from target attributes and as such does not call the bmc
for the type information.
* Added the new IpmiConfigLookup class in the IPMI namespace.
  This class provides static methods to retrive sensor data
  provided by the IPMI_SENSOR target attribute. In addition to
  sensor type, the sensor entity id and sensor name can be
  obtined from this class based upon a sensor number.
* Changed sendErrLogToBMC to call IpmiConfigLookup::getSensorType
  instead of Sensor::SensorBase::getSensorInfo.

RTC: 164493
Change-Id: Ie8e7fa5c37fa972d6cecd1e4b69e2d0cb7504533
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42128
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>