Change Bootloader data alignments
The data defined in bootloader_data.H is not aligned as originally
thought.  Some changes should be made to this file and the
BlData.pm tool to nicely align this data and output it.
Also add static_assert statements to alert for future changes.
Use default address to find data and traces with optional argument
to override that address.

Change-Id: I4ed61273baa54b2c7cb510ef9a24cebb239d9300
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42057
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Stephen M. Cprek <smcprek@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>