Removed i2cSetupActiveMasters from last the IPL Step.
The i2cSetupActiveMasters function was called during the last ipl
step in order to set the i2c bus speed to 400khz. This
functionality was determined to be redundant as the OPAL/PHYP code
initializes i2c bus speed. The call to i2cSetupActiveMasters has
been commented out for over a year.

RTC: 164492
Change-Id: I9c6a77234b15c0837768fc30e2a4f4a4da05023c
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42290
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>