kernel: Fix cmp in shutdown.S for modern binutils
This is a further fix that was missed in 10983d35f568 ("Convert cmpi to
cmpwi to fix build for modern binutils").

In the past binutils would assume 'cmp' meant 'cmpw'. Modern binutils
require the explict use of cmpw or cmpd.

The first two instances are dealing with SPRs, which are 32-bit, so cmpw
is correct.

It is not clear what to do in the third use of cmp, but given old toolchains
have generated cmpw lets assume that we should maintain the behaviour.

Change-Id: Iee5dd5903dcd7ac4028bab0176e08ce3db23b2d5
Signed-off-by: Joel Stanley <joel.stanley@au1.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/42476
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>